refactor: move holograms into types directory
