+++ paper/src/main/java/com/badbones69/crazyenvoys/api/CrazyManager.java
@@ -22,9 +22,9 @@ import com.badbones69.crazyenvoys.api.objects.FlareSettings;
 import com.badbones69.crazyenvoys.api.objects.LocationSettings;
 import com.badbones69.crazyenvoys.api.objects.misc.Tier;
 import com.badbones69.crazyenvoys.listeners.timer.CountdownTimer;
-import com.badbones69.crazyenvoys.support.holograms.CMIHologramsSupport;
+import com.badbones69.crazyenvoys.support.holograms.types.CMIHologramsSupport;
 import com.badbones69.crazyenvoys.support.claims.WorldGuardSupport;
-import com.badbones69.crazyenvoys.support.holograms.DecentHologramsSupport;
+import com.badbones69.crazyenvoys.support.holograms.types.DecentHologramsSupport;
 import com.ryderbelserion.vital.common.configuration.YamlFile;
 import com.ryderbelserion.vital.common.configuration.YamlManager;
 import com.ryderbelserion.vital.common.configuration.objects.CustomFile;
